Add unit tests for PathTracker class
Add a very basic unit test and turn unit testing on in tox now that
there is something for it to do. Running "coverage" with zero unit
tests would have caused it to fail which is why it was commented
out before.

Obviously these unit tests are just a starting point and cover very
little of value in real terms.