Export a "connect" function from the binding
Having the auth code inside the __init__ method of the Opsramp
class makes it difficult to unit test. Change to using a function
as the external interface of this module instead, so that it'll
be easier to add unit tests of the classes later because we'll be
able to instantiate the Opsramp class without needing an actual
web service to which to authenticate.