Accept XML errors and use ApiObject in connect()
Many OpsRamp API calls return errors as XML which currently get lost
because we don't accept that return type. All the user sees is a 406
HTTP error which is not very helpful. Modify all our calls to accept
XML response optionally.

While here I noticed that the connect() function could be implemented
more simply using an ApiObject instance so modify it to do that.