Add "integration" section of the API
First steps into the world of "integrations" in OpsRamp.
I'm also including an examples update to exercise it.

One thing worth noting is that we found some of the names of the
OpsRamp "integration" API endpoints confusing, in that it took us a
while to figure out what the returned data sets actually meant.
We concluded that those endpoint names were confusing and so we
decided to use more obvious names here. In particular "types" and
"instances" rather than "available" and "installed".