Include inbound auth type in the Custom integration
Tests show that supplying the auth type up front causes OpsRamp
to return the actual OAUTH2 key & secret in its response. We will
need those to use the integration so provide access to that method.

Redact them from the output of examples.py in case they end up in
logs or bug reports.