Add "samples" subdir with standalone example scripts
The main examples.py has got large so while it's still good for a smoke
test it's not so good for conveying how the library is supposed to be
used. Add some new single-purpose samples to show that better.

I have written these to output JSON because that might make them useful
in the field. At the same time I know I'm taking a chance on people not
understanding that this is only one possible use of the return values.