Major refactor to flatten the object hierarchy
Analysis of the design has shown that we are not getting value from
the extra layer of classes representing individual instances of
clients, scripts, integrations etc. This change flattens the
hierarchy by removing them and pushing their functionality up a
level into functions of the collection classes that now take an
optional "uuid" parameter to select a specific instance.