requests_mock for unit test coverage on base.py
This file contains the fundamental base classes of this module so
it's important to have good unit test coverage on them.

These are the first unit tests that result in actual attempts to
call the OpsRamp API, so I'm using the requests_mock module to stub
those out and control the return values.