Add coveralls to CI run. (#22)
Add coveralls to CI run.

If the appropriate env vars are present then run coveralls

The SaaS coveralls.io is offline at the moment and that triggers me to
add code to continue if it reports an error. We don't want to fail the
merge for that, so let it go to stderr but ignore the result.