Add CircleCI config file
This is a first attempt and will need tweaking but it seems to work for now.
It's only testing on Python3 which is not ideal; would like to test 2.7 also.
That can be added as a separate commit once I figure out how.

Remove the Travis CI config file because we're moving to CircleCI and don't
need both.