Add roles to README and update class diagram
Turn the class diagram on its side because it's getting wider and wider
not taller and taller so this will scale better.

This commit also includes some changes to examples.py to display the
roles and also clean up some other output by printing it in YAML
format which is easier to parse by eye.