YAML print of create email integration result
It appears that OpsRamp has a bug whereby the email address for the
new email integration is not returned by the CREATE call. It is
deliberately omitted from GET so this bug means it can't be retrieved
programmatically at all.

This change to the relevant sample makes it print out the entire
return struct in YAML so that it's easy to see the omission and
will be easy to detect when it gets fixed in the future.