Move to Python 3.7 for CI and Docker
I am deliberately leaving tox.ini as "py3" which means "the latest version
of Python3 that is available on this machine" so that we are not mandating
the version that the committer must install on their workstation. The CI
checks will run on the exact version anyway so we don't need to worry.