requests returns a different exception type on JSON errors
Field experience has shown that the Python requests package
returns a different exception type on JSON parsing errors
than originally thought.

Interestingly, this was discovered by running opsramp.examples
against a real OpsRamp server. The text response returned by
the "get agent script" API call exposed this error because the
attempt to parse it as JSON failed as expected, but the actual
type of exception raised was different from the one the unit
tests are using.