Expand the resources class.
Add search and other mainstream functions to the Resources class.
I deliberately have not covered everything that is listed on the
"resources" section of the OpsRamp docs because imho some of that
does not belong in this class e.g. they mix resource *types* with
resource instances, which are not the same thing.