Resources sometimes a list, sometimes not.
Amusingly, tests at partner level show that sometimes it returns a
proper paginated result and sometimes it doesn't. Possibly related
to the length of the list I don't know. Anyway, this is a change to
check if the result is already in the "normal" format and if so to
return it verbatim.