Nitpicks from flake8-bugbear
These are mostly around the use of empty dicts {} as defaults in function
definitions. Technically that syntax results in a static dict that is used
again every time a caller omits the argument, so in theory if the function
made changes to the dict those would persist to the next call. I don't
see any instances of us doing that, but I'm removing the possibility.