Add newer tenant properties to examples.py
Some of the most recently added properties are not being exercised by
examples.py so fix that. As a driveby, the timeout value for the 429
rate limiting code is too short so increase it. Running examples.py
nearly always causes rate limiting because it makes so many API calls
in a short period, so this is a good test of the retry code and in
fact it was getting retries exceeded which is why I increased the values.