Move search() into the base class ORapi
Many classes have their own "search" functions but most of them are
identical so it makes sense to move into the base class and override
the minority that need to behave differently.

Change the unit tests to accommodate that, but also refactor some of
them to avoid reusing the requests_mock object because that can hide
issues where a test is succeeding only because a previous one has
primed the mock for it.