Top-level API support for the Metrics tree.
The OpsRamp metrics API is spread all over the place. This is a simple
initial class to expose the entire API as a top-level object directly
accessed from the "binding" object.

You can access any GET method of that tree using this object and an
appropriate pattern. Some examples are given in the "sample" app.

I would hope to add a more usable representation of "metric type" and
time series metric values in the future, but it's complex because the
OpsRamp metrics API is very oddly structured.