Automated retry on POST for rate limiting
OpsRamp have introduced new rate limits that require us to retune the
handler for the HTTP 429 rate limiting status. In particular the limit
for POST is very low (~2 per second) which means that any configuration
code that is creating multiple objects is very likely to hit it. To
help with this we are introducing retries on POST if and only if the
error was rate limiting. This is on an assumption that 429 means the
POST did not happen and therefore it's safe to retry it.