Remove the need for urllib3 1.26+
This is a very new version of urllib3 and even Ubuntu 20.04 doesn't
provide it by default. We're only using the DEFAULT_ALLOWED_METHODS
attribute and we can easily fallback to an older attribute that was
present in earlier urllib3 versions. The older one is deprecated and
will go away in the future, so I am leaving the new name in place
as the first choice if it exists.