Make unit test URLs mocks case-sensitive.
OpsRamp URLs are case-sensitive so make the mocks in our unit tests
check for that. By default request_mock is case-insensitive and we
have to explicitly enable that.

conftest.py is the canonical filename that pytest reads at the start
of each run to provide a place to add hooks, so use it to implement
this.