Add "session" property to the ApiWrapper class
Allow the user to supply a customized requests session if they need to.
This is an advanced use case because in general we would not want the
caller to have to deal with the requests module directly; it's effectively
an implementation detail. However session objects do sometimes need to be
built specifically the use case in hand, so we should allow that.