Add --debug support to the samples
Use argparse to implement a --debug flag that enables Python DEBUG
logging which can be very useful as a learning tool for what is
going on under the hood, which is one of the points of these samples.

I have cut/paste the code into each one because I want each sample
file to stand alone and not have any dependencies.

While here, fix a few problems with the main() function in the
service map examples to make them match the others.