OK is not the only HTTP success code.
Anything in the range 200-299 is a successful result and in particular
POST operations often return http CREATED or NO_CONTENT if there is
nothing more to communicate except that the POST succeeded.

Also refactor the unit tests in this area slightly, they are a bit
old-fashioned in having long functions instead of several short ones
which are easier to spot in the output if they fail.