Move away from deprecated CircleCI images
https://discuss.circleci.com/t/legacy-convenience-image-deprecation
CircleCI have changed the naming pattern for their images.

As part of this commit I am also removing the caching of build env
from previous builds. It causes problems when changing images like
this because the cache is not invalidated and the old contents are
not appropriate to the new image. There's no real need to cache
anyway because even the complete build is quite fast.