Explicitly include requirements.txt in the sdist
Add MANIFEST.in to explicitly include requirements.txt in the sdist.

That file is definitely needed by setup.py but there is no explicit
instruction to add it to the source distro in the existing codebase.
It's not clear why it's being added at the moment on most machines
regardless. But I would prefer to make it explicit anyway.