Run our unit tests using GitHub Action script (#124)
Try both variations of our build. I would prefer to start with a clean slate
for each one but I don't know how to do that yet so serial will have to do.

This repo has a script that is used to run the tests so use that directly
like the other CI configurations already do.