We are losing access to CircleCI so remove it.
The account on which this CI was running is not beinge renewed so turn off
CircleCI by removing the config file. We can resurrect it later if/when we
get access to a different account under which to run.

We have very similar CI running in GitHub actions anyway so it's not the
end of the world; our tests will still run.