Fix articlesList bug, update sample
I don't know if this changed on the OpsRamp side, or just has never
worked. Either way "articles/search" is not correct it needs to be
"articlesList". Fix that bug + unit test.

Also take the opportunity to make the sample that shows some basic
information about the knowledge base work a bit better and show in
particular how to reach the article content (which seems to be in
HTML format though that's not stated in the OpsRamp API docs).